requirements.txt
discord.py
nest_asyncio
email-validator
python-dotenv